requirements.txt
paramiko
langchain
openai
streamlit
tiktoken
pillow